refactor: 将各类代码放到合适的地方去了
