+++ .gitignore
@@ -32,3 +32,4 @@ coco_caption/pyciderevalcap/tokenizer/tmpzNW4I2
 coco_caption/pycocoevalcap/tokenizer/tmp5zqynfzy
 coco_caption/pycocoevalcap/tokenizer/tmplcjcdusq
 coco_caption/pycocoevalcap/tokenizer/tmps82e821g
+/outdated/
